environment.py: more defensive check before rm -rf
Currently `old_root` is computed by reading the symlink at `self.root`.
We should be more defensive in removing it by checking that it is in the
same directory as the new root. Otherwise, in the worst case, when
someone runs `spack env create --with-view=./view -d .` and `view`
already exists and is a symlink to `/`, Spack effectively runs `rm -rf /`.

+++ lib/spack/spack/environment/environment.py
@@ -541,8 +541,12 @@ class ViewDescriptor(object):
             raise SpackEnvironmentViewError(msg)
         rename(tmp_symlink_name, self.root)
 
-        # remove old_root
-        if old_root and os.path.exists(old_root):
+        # remove old_root when it looks like an old root
+        if (
+            old_root and
+            os.path.exists(old_root) and
+            os.path.samefile(os.path.dirname(new_root), os.path.dirname(old_root))
+        ):
             try:
                 shutil.rmtree(old_root)
             except (IOError, OSError) as e:
